Work on medmeapp importer
